# @FIR-996: Fix and issue where the hugging face model name was not