normalized and add details to server busy message
This change adds following
1. Server busy message adds details on what is holding the server busy
2. Handle the scenarios where the model name does not have version so a ":latest" is added by ollama when downloaded from hugging face. This was handled for model uploaded manually but not for hugging face download

the test results are as follows
WARNING: CORS_ALLOW_ORIGIN IS SET TO '*' - NOT RECOMMENDED FOR PRODUCTION DEPLOYMENTS.

INFO  [open_webui.env] VECTOR_DB: chroma
INFO  [open_webui.env] Embedding model set: sentence-transformers/all-MiniLM-L6-v2 /proj/work/atrivedi/workspace/09_30_2025/open-webui/.venv/lib/python3.12/site-packages/pydub/utils.py:300: SyntaxWarning: invalid escape sequence '\('
  m = re.match('([su]([0-9]{1,2})p?) \(([0-9]{1,2}) bit\)$', token)
/proj/work/atrivedi/workspace/09_30_2025/open-webui/.venv/lib/python3.12/site-packages/pydub/utils.py:301: SyntaxWarning: invalid escape sequence '\('
  m2 = re.match('([su]([0-9]{1,2})p?)( \(default\))?$', token)
/proj/work/atrivedi/workspace/09_30_2025/open-webui/.venv/lib/python3.12/site-packages/pydub/utils.py:310: SyntaxWarning: invalid escape sequence '\('
  elif re.match('(flt)p?( \(default\))?$', token):
/proj/work/atrivedi/workspace/09_30_2025/open-webui/.venv/lib/python3.12/site-packages/pydub/utils.py:314: SyntaxWarning: invalid escape sequence '\('
  elif re.match('(dbl)p?( \(default\))?$', token):
/proj/work/atrivedi/workspace/09_30_2025/open-webui/.venv/lib/python3.12/site-packages/pydub/utils.py:170: RuntimeWarning: Couldn't find ffmpeg or avconv - defaulting to ffmpeg, but may not work
  warn("Couldn't find ffmpeg or avconv - defaulting to ffmpeg, but may not work", RuntimeWarning)
WARNI [langchain_community.utils.user_agent] USER_AGENT environment variable not set, consider setting it to identify your requests. Fetching 30 files: 100%|██████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 30/30 [00:00<00:00, 15984.39it/s]
INFO:     Started server process [33196]
INFO:     Waiting for application startup.
127.0.0.1 - - [30/Sep/2025 21:38:53] "POST /api/receive HTTP/1.1" 200 -
127.0.0.1 - - [30/Sep/2025 21:45:26] "POST /api/system-info HTTP/1.1" 200 -
127.0.0.1 - - [30/Sep/2025 21:45:46] "POST /api/health-check HTTP/1.1" 200 -

Flask Logs
			   T_CommonMode=0us LTR1.2_Threshold=32768ns
		L1SubCtl2: T_PwrOn=10us
	Capabilities: [910 v1] Data Link Feature <?>
	Capabilities: [920 v1] Lane Margining at the Receiver <?>
	Capabilities: [9c0 v1] Physical Layer 16.0 GT/s <?>
	Kernel driver in use: nvme
	Kernel modules: nvme

Setting PCI command bit for device: 0000:04:00.0
sudo ./copy2fpga-x86 /var/snap/ollama/common/models/blobs/sha256-94c6ddc5bf5a2c0ef1130aac48859b0eaf33af33a6aa1527d8fd7354ea0f5206 PCIE base address is: 0xfcf0300000

BEGIN FILE TRANSFER (Send process):
====================================
On FPGA Target, Start Receive process.. : File size to be transfered: 270885792 bytes

Starting Transfer >>>>>>
Current time: 21:38:31
Current date: 2025-09-30

Time taken for transfer: 17.858510 seconds
Current time: 21:38:49
Current date: 2025-09-30

File Transfer completed.
======================================
 Running on system: fpga3
======================================
Valid system.. continuing setup
Found PCI device at BDF: 0000:04:00.0
Removing PCI device: 0000:04:00.0
Rescanning PCI bus...
Dumping PCI data using script: /aws/proj/rel/sw/platform/release_v0.1.1.tsv026_04_15_2025/scripts/dump-pci.sh setpci -s1:0 VENDOR_ID
15b7
 setpci -s1:0 DEVICE_ID
5017
 setpci -s1:0 REVISION
02
setpci -s1:0 CLASS_DEVICE
0108
setpci -s1:0 SUBSYSTEM_VENDOR_ID
15b7
 setpci -s1:0 SUBSYSTEM_ID
5017
 setpci -s1:0 BASE_ADDRESS_0
f6f00004
setpci -s1:0 BASE_ADDRESS_1
00000000
setpci -s1:0 BASE_ADDRESS_2
00000000
setpci -s1:0 BASE_ADDRESS_3
00000000
lspci -vvvnns1:0
01:00.0 Non-Volatile memory controller [0108]: Sandisk Corp Device [15b7:5017] (rev 01) (prog-if 02 [NVM Express])
	Subsystem: Sandisk Corp Device [15b7:5017]
	Control: I/O+ Mem+ BusMaster+ SpecCycle- MemWINV- VGASnoop- ParErr- Stepping- SERR- FastB2B- DisINTx+
	Status: Cap+ 66MHz- UDF- FastB2B- ParErr- DEVSEL=fast >TAbort- <TAbort- <MAbort- >SERR- <PERR- INTx-
	Latency: 0, Cache Line Size: 64 bytes
	Interrupt: pin A routed to IRQ 49
	IOMMU group: 12
	Region 0: Memory at f6f00000 (64-bit, non-prefetchable) [size=16K]
	Capabilities: [80] Power Management version 3
		Flags: PMEClk- DSI- D1- D2- AuxCurrent=0mA PME(D0-,D1-,D2-,D3hot-,D3cold-)
		Status: D0 NoSoftRst+ PME-Enable- DSel=0 DScale=0 PME-
	Capabilities: [90] MSI: Enable- Count=1/32 Maskable- 64bit+
		Address: 0000000000000000  Data: 0000
	Capabilities: [b0] MSI-X: Enable+ Count=65 Masked-
		Vector table: BAR=0 offset=00003000
		PBA: BAR=0 offset=00002000
	Capabilities: [c0] Express (v2) Endpoint, MSI 00
		DevCap:	MaxPayload 512 bytes, PhantFunc 0, Latency L0s <1us, L1 unlimited
			ExtTag- AttnBtn- AttnInd- PwrInd- RBE+ FLReset+ SlotPowerLimit 0.000W
		DevCtl:	CorrErr+ NonFatalErr+ FatalErr+ UnsupReq+
			RlxdOrd+ ExtTag- PhantFunc- AuxPwr- NoSnoop+ FLReset-
			MaxPayload 512 bytes, MaxReadReq 512 bytes
		DevSta:	CorrErr+ NonFatalErr- FatalErr- UnsupReq+ AuxPwr- TransPend-
		LnkCap:	Port #0, Speed 16GT/s, Width x4, ASPM L1, Exit Latency L1 <8us
			ClockPM+ Surprise- LLActRep- BwNot- ASPMOptComp+
		LnkCtl:	ASPM Disabled; RCB 64 bytes, Disabled- CommClk+
			ExtSynch- ClockPM- AutWidDis- BWInt- AutBWInt-
		LnkSta:	Speed 16GT/s (ok), Width x4 (ok)
			TrErr- Train- SlotClk+ DLActive- BWMgmt- ABWMgmt-
		DevCap2: Completion Timeout: Range B, TimeoutDis+ NROPrPrP- LTR+
			 10BitTagComp+ 10BitTagReq- OBFF Not Supported, ExtFmt+ EETLPPrefix-
			 EmergencyPowerReduction Not Supported, EmergencyPowerReductionInit-
			 FRS- TPHComp- ExtTPHComp-
			 AtomicOpsCap: 32bit- 64bit- 128bitCAS-
		DevCtl2: Completion Timeout: 50us to 50ms, TimeoutDis- LTR+ OBFF Disabled,
			 AtomicOpsCtl: ReqEn-
		LnkCap2: Supported Link Speeds: 2.5-16GT/s, Crosslink- Retimer+ 2Retimers+ DRS-
		LnkCtl2: Target Link Speed: 16GT/s, EnterCompliance- SpeedDis-
			 Transmit Margin: Normal Operating Range, EnterModifiedCompliance- ComplianceSOS-
			 Compliance De-emphasis: -6dB
		LnkSta2: Current De-emphasis Level: -3.5dB, EqualizationComplete+ EqualizationPhase1+
			 EqualizationPhase2+ EqualizationPhase3+ LinkEqualizationRequest-
			 Retimer- 2Retimers- CrosslinkRes: unsupported
	Capabilities: [100 v2] Advanced Error Reporting
		UESta:	DLP- SDES- TLP- FCP- CmpltTO- CmpltAbrt- UnxCmplt- RxOF- MalfTLP- ECRC- UnsupReq- ACSViol-
		UEMsk:	DLP- SDES- TLP- FCP- CmpltTO- CmpltAbrt- UnxCmplt- RxOF- MalfTLP- ECRC- UnsupReq- ACSViol-
		UESvrt:	DLP+ SDES- TLP- FCP+ CmpltTO- CmpltAbrt- UnxCmplt- RxOF+ MalfTLP+ ECRC- UnsupReq- ACSViol-
		CESta:	RxErr- BadTLP- BadDLLP- Rollover- Timeout- AdvNonFatalErr-
		CEMsk:	RxErr- BadTLP- BadDLLP- Rollover- Timeout- AdvNonFatalErr+
		AERCap:	First Error Pointer: 00, ECRCGenCap+ ECRCGenEn- ECRCChkCap+ ECRCChkEn-
			MultHdrRecCap- MultHdrRecEn- TLPPfxPres- HdrLogCap-
		HeaderLog: 00000000 00000000 00000000 00000000
	Capabilities: [1b8 v1] Latency Tolerance Reporting
		Max snoop latency: 1048576ns
		Max no snoop latency: 1048576ns
	Capabilities: [300 v1] Secondary PCI Express
		LnkCtl3: LnkEquIntrruptEn- PerformEqu-
		LaneErrStat: 0
	Capabilities: [900 v1] L1 PM Substates
		L1SubCap: PCI-PM_L1.2+ PCI-PM_L1.1+ ASPM_L1.2+ ASPM_L1.1- L1_PM_Substates+
		L1SubCtl1: PCI-PM_L1.2- PCI-PM_L1.1- ASPM_L1.2- ASPM_L1.1-
			   T_CommonMode=0us LTR1.2_Threshold=32768ns
		L1SubCtl2: T_PwrOn=10us
	Capabilities: [910 v1] Data Link Feature <?>
	Capabilities: [920 v1] Lane Margining at the Receiver <?>
	Capabilities: [9c0 v1] Physical Layer 16.0 GT/s <?>
	Kernel driver in use: nvme
	Kernel modules: nvme

Setting PCI command bit for device: 0000:04:00.0
sudo ./copy2fpga-x86 /var/snap/ollama/common/models/blobs/sha256-3ec6f8995cf5d8d17d9e2e013b5439c3b807e57d3d0d4c9d27399b602ce3ccf2

 PortCommonModeRestoreTime=32us [PortTPowerOnTime=10us](
<img width="1468" height="806" alt="Screenshot 2025-09-30 at 8 43 02 PM" src="https://github.com/user-attachments/assets/1fd970a6-1a4b-4ada-8460-6b0297dd3375" />
<img width="1352" height="688" alt="Screenshot 2025-09-30 at 9 45 40 PM" src="https://github.com/user-attachments/assets/d5a1c43d-0979-4123-975c-1c89f95098c2" />
<img width="1873" height="324" alt="Screenshot 2025-09-30 at 9 39 20 PM" src="https://github.com/user-attachments/assets/8e2fcd76-d8ef-4852-852f-c4d7ffe74ebd" />
<img width="1361" height="696" alt="Screenshot 2025-09-30 at 9 21 54 PM" src="https://github.com/user-attachments/assets/3d72408b-f5d9-4abe-8b12-4fb087616b22" />
<img width="1758" height="203" alt="Screenshot 2025-09-30 at 9 21 18 PM" src="https://github.com/user-attachments/assets/5341c4ba-e30d-4f8f-9423-6ecd62c3d427" />
<img width="1888" height="805" alt="Screenshot 2025-09-30 at 8 43 46 PM" src="https://github.com/user-attachments/assets/fac0a392-cd9c-4348-8e92-0ce17d90d6ea" />
<img width="1891" height="828" alt="Screenshot 2025-09-30 at 8 46 51 PM" src="https://github.com/user-attachments/assets/dfe1083a-86ff-47fb-9380-ee04c003cdc2" />
<img width="1468" height="805" alt="Screenshot 2025-09-30 at 8 48 35 PM" src="https://github.com/user-attachments/assets/e86aecdd-60e1-45ad-bff8-978f7af9e391" />
<img width="1463" height="794" alt="Screenshot 2025-09-30 at 8 44 26 PM" src="https://github.com/user-attachments/assets/dddd677e-ef55-4266-9f68-aac40241a9c2" />
<img width="1221" height="617" alt="Screenshot 2025-09-30 at 8 42 20 PM" src="https://github.com/user-attachments/assets/e26aa41e-8f92-465d-bcca-3a551b3dee36" />
<img width="1325" height="709" alt="Screenshot 2025-09-30 at 9 45 59 PM" src="https://github.com/user-attachments/assets/a318ac59-78b2-48e8-9957-66d57cb718a0" />)